[TODO]: Have SVGame adjust the encoded animationIDs based on other client pmove events, or perhaps just... use the entity events system here as well. [SKM]: Player animation's seem quite decent for now. So we'll stick to this. [Messaging]: No more OLD_FRAME being sent around. [Messaging]: Properly WriteUintBase128 for entity frames. (It was before so decoded to a short, oops.)

+++ inc/common/messaging.h
@@ -79,8 +79,8 @@ extern "C" {
 		uint32_t    renderfx;
 		uint32_t    effects;
 
-		uint16_t    frame;
-		uint16_t	old_frame;
+		uint32_t    frame;
+		uint32_t	old_frame;
 
 		uint16_t    sound;
 		uint8_t     event;
